+++ test/unit/fuzz-tools.test.js
@@ -236,6 +236,7 @@ import * as entsleepV243 from '../../lib/entsleep-v243.js';
 import * as sportsmskV244 from '../../lib/sportsmsk-v244.js';
 import * as hemedermV245 from '../../lib/hemederm-v245.js';
 import * as ibdV246 from '../../lib/ibd-v246.js';
+import * as pedstoxV247 from '../../lib/pedstox-v247.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -416,6 +417,7 @@ const MODULES = {
   'sportsmsk-v244.js': sportsmskV244,
   'hemederm-v245.js': hemedermV245,
   'ibd-v246.js': ibdV246,
+  'pedstox-v247.js': pedstoxV247,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];